✨ Dynamic routing for student building pages
